[refactor] scroll 관련 훅 분리
